api/requirements.txt

flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.1
mysqlclient==2.2.1
marshmallow==3.20.1
PyJWT==2.8.0
